source/eksfedctl/requirements.txt
boto3
pyyaml
requests